修正优先级计划的实时回退索引
保留原始重试位置用于新候选集合，并整合回归测试及补充三数据库验证。
